вынести моковые данные в константы, выполнить рефакторинг функций, генерирующих моковые данные, вынести шаблоны в отдельные модули
