+++ src/utils.js
@@ -1,3 +1,8 @@
+import dayjs from 'dayjs';
+import duration from 'dayjs/plugin/duration';
+
+dayjs.extend(duration);
+
 const convertMinutesToHours = (minutesDuration) => {
   const hours = Math.floor(minutesDuration / 60);
   const minutes = Math.floor(minutesDuration % 60);
@@ -12,4 +17,22 @@ const getRandomInteger = (a = 0, b = 1) => {
   return Math.floor(lower + Math.random() * (upper - lower + 1));
 };
 
-export { convertMinutesToHours, getRandomInteger };
+const getRandomItem = (items) => items[getRandomInteger(0, items.length - 1)];
+
+const formatISOStringToDate = (ISOString) => dayjs(ISOString).format('DD MMMM YYYY');
+
+const formatISOStringToDateWithTime = (ISOString) => dayjs(ISOString).format('YYYY/MM/DD HH:MM');
+
+const formatISOStringToYear = (ISOString) => dayjs(ISOString).format('YYYY');
+
+const formatMinutesToTime = (minutes) => dayjs.duration(minutes, 'minutes').format('H[h] mm[m]');
+
+export {
+  convertMinutesToHours,
+  getRandomInteger,
+  getRandomItem,
+  formatISOStringToDate,
+  formatISOStringToYear,
+  formatMinutesToTime,
+  formatISOStringToDateWithTime,
+};
