+++ src/view/film-details-comment-view.js
@@ -1,25 +1,6 @@
 import { createElement } from '../render.js';
-import dayjs from 'dayjs';
 
-const createFilmDetailsCommentTemplate = (commentObj) => {
-  const { author, comment, date, emotion } = commentObj;
-
-  return `
-    <li class="film-details__comment">
-      <span class="film-details__comment-emoji">
-      <img src="./images/emoji/${emotion}.png" width="55" height="55" alt="emoji-smile">
-      </span>
-      <div>
-        <p class="film-details__comment-text">${comment}</p>
-        <p class="film-details__comment-info">
-          <span class="film-details__comment-author">${author}</span>
-          <span class="film-details__comment-day">${dayjs(date).format('YYYY/MM/DD HH:MM')}</span>
-          <button class="film-details__comment-delete">Delete</button>
-        </p>
-      </div>
-    </li>`;
-};
-
+import { createFilmDetailsCommentTemplate } from './film-details-comment-template.js';
 export default class FilmDetailsCommentView {
   constructor(comment) {
     this.comment = comment;
